+++ bzt/engine.py
@@ -29,7 +29,6 @@ from collections import namedtuple, defaultdict
 from distutils.version import LooseVersion
 from json import encoder
 
-import psutil
 import yaml
 from yaml.representer import SafeRepresenter
 
@@ -72,9 +71,7 @@ class Engine(object):
         self.interrupted = False
         self.check_interval = 1
         self.stopping_reason = None
-        self.__disk_counters = None
-        self.__net_counters = None
-        self.__counters_ts = None
+        self.engine_loop = 0
 
     def configure(self, user_configs, read_config_files=True):
         """
@@ -164,6 +161,7 @@ class Engine(object):
             now = time.time()
             diff = now - prev
             delay = self.check_interval - diff
+            self.engine_loop_percent = diff / self.check_interval
             self.log.debug("Iteration took %.3f sec, sleeping for %.3f sec...", diff, delay)
             if delay > 0:
                 time.sleep(delay)
@@ -476,50 +474,6 @@ class Engine(object):
         else:
             self.aggregator = self.instantiate_module(cls)
         self.aggregator.prepare()
-
-    def engine_resource_stats(self):
-        """
-        Get local resource stats
-
-        :return: namedtuple
-        """
-        stats = namedtuple("ResourceStats", ('cpu', 'disk_usage', 'mem_usage',
-                                             'rx', 'tx', 'dru', 'dwu'))
-        rx_bytes, tx_bytes, dru, dwu = self.__get_resource_stats()
-        # TODO: measure and report check loop utilization
-        return stats(
-                cpu=psutil.cpu_percent(),
-                disk_usage=psutil.disk_usage(self.artifacts_dir).percent,
-                mem_usage=psutil.virtual_memory().percent,
-                rx=rx_bytes, tx=tx_bytes, dru=dru, dwu=dwu
-        )
-
-    def __get_resource_stats(self):
-        """
-        Get network and disk counters
-        :return: tuple
-        """
-        if not self.__counters_ts:
-            self.__disk_counters = psutil.disk_io_counters()
-            self.__net_counters = psutil.net_io_counters()
-            self.__counters_ts = datetime.datetime.now()
-            time.sleep(0.2)  # small enough for human, big enough for machine
-
-        now = datetime.datetime.now()
-        interval = (now - self.__counters_ts).total_seconds()
-
-        net = psutil.net_io_counters()
-        tx_bytes = (net.bytes_sent - self.__net_counters.bytes_sent) / interval
-        rx_bytes = (net.bytes_recv - self.__net_counters.bytes_recv) / interval
-        self.__net_counters = net
-
-        disk = psutil.disk_io_counters()
-        dru = (disk.read_bytes - self.__disk_counters.read_bytes) / interval
-        dwu = (disk.write_bytes - self.__disk_counters.write_bytes) / interval
-        self.__disk_counters = disk
-
-        self.__counters_ts = now
-        return rx_bytes, tx_bytes, dru, dwu
 
     def _set_up_proxy(self):
         proxy_settings = self.config.get("settings").get("proxy")