+++ site/dat/docs/Changelog.md
@@ -3,6 +3,7 @@
 ## 1.1.1 (next)
  - maximize browser window in Selenium test, when possible
  - add graphite support to monitoring service
+ - add local monitoring service
  - create [Docker](Docker.md) image with Taurus inside
 
 ## 1.1.0